Day 3
This was day 3 of my coding a mini project a day everyday for my whole first semester to learn programming 🎉🎉🎉

I programmed a random number generator guessing game from scratch and didn't even use external sources so that's a big win for me i guess hahahaha! I'm proud of what I did as of now and I'm excited for the future of my programming journey.